requirements.txt
httpx
pydantic
exa-py
fastapi
requests
uvicorn
agno
python-dotenv
typing-extensions
a2a-sdk[sqlite]
sqlalchemy
google-genai
uagents-adapter
